#29 moved slide image upload from db to file system

+++ web-app/Controllers/SlideController.cs
@@ -1,10 +1,10 @@
 ﻿using BAL.Interfaces;
-using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
-using Newtonsoft.Json;
-using System.Linq;
+using Microsoft.Extensions.Hosting;
+using System;
+using System.IO;
 using System.Threading.Tasks;
-using web_app.ViewModels.Presentation;
 using web_app.ViewModels.Slide;
 
 namespace web_app.Controllers
@@ -12,10 +12,12 @@ namespace web_app.Controllers
     public class SlideController : Controller
     {
         private readonly ISlideService slideService;
+        private readonly IWebHostEnvironment hostEnvironment;
 
-        public SlideController(ISlideService slideService)
+        public SlideController(ISlideService slideService, IWebHostEnvironment hostEnvironment)
         {
             this.slideService = slideService;
+            this.hostEnvironment = hostEnvironment;
         }
 
         [HttpPost]
@@ -23,22 +25,43 @@ namespace web_app.Controllers
         {
             var slide = await this.slideService.AddTitleSlide(presentationId);
 
+            if(slide is null)
+            {
+                return new JsonResult(NotFound());
+            }
+
             var slideViewModel = new SlideViewModel
             {
                 Id = slide.Id,
                 Title = slide.Title,
                 Text = slide.Text,
-                Image = slide.Image,
+                Image = slide.Image
             };
 
             return new JsonResult(slideViewModel);
         }
 
         [HttpPost]
-        public async Task<JsonResult> AddImageSlide([FromForm] IFormFile image, [FromForm] string json)
+        public async Task<JsonResult> AddImageSlide([FromForm] SlideImageViewModel viewModel)
         {
-            var presentation = JsonConvert.DeserializeObject<PresentationNameViewModel>(json);
-            var slide = await this.slideService.AddImageSlide(presentation.Id, image);
+            if (viewModel.Image == null && viewModel.Image.ContentType.StartsWith("image"))
+            {
+                return new JsonResult(BadRequest());
+            }
+
+            string wwwroot = hostEnvironment.WebRootPath;
+            string fileName = Guid.NewGuid().ToString();
+            var uploads = Path.Combine(wwwroot, @"images\slides");
+            var extension = Path.GetExtension(viewModel.Image.FileName);
+
+            using (var fileStream = new FileStream(Path.Combine(uploads, fileName + extension), FileMode.Create))
+            {
+                viewModel.Image.CopyTo(fileStream);
+            }
+
+            var imagePath = @"\images\slides\" + fileName + extension;
+
+            var slide = await this.slideService.AddImageSlide(viewModel.PresentationId, imagePath);
 
             var slideViewModel = new SlideViewModel
             {
@@ -49,20 +72,67 @@ namespace web_app.Controllers
             return new JsonResult(slideViewModel);
         }
 
-        [HttpDelete]
-        public async Task<JsonResult> Remove([FromQuery] int id)
+      
+        [HttpPut]
+        public async Task<JsonResult> EditTitle([FromBody] SlideTitleViewModel viewModel)
         {
-            await this.slideService.Remove(id);
+            var result = await this.slideService.EditTitle(viewModel.Id, viewModel.Title);
+
+            if(result is false)
+            {
+                return new JsonResult(NotFound());
+            }
 
             return new JsonResult(Ok());
         }
 
         [HttpPut]
-        public async Task<JsonResult> Edit([FromBody] SlideViewModel viewModel)
+        public async Task<JsonResult> EditText([FromBody] SlideTextViewModel viewModel)
         {
-            await this.slideService.Edit(viewModel.Id, viewModel.Title, viewModel.Text);
+            var result = await this.slideService.EditText(viewModel.Id, viewModel.Text);
+
+            if (result is false)
+            {
+                return new JsonResult(NotFound());
+            }
 
             return new JsonResult(Ok());
         }
+
+        [HttpPut]
+        public async Task<JsonResult> EditOnDragAndDrop([FromBody] SlideDragAndDropViewModel viewModel)
+        {
+            var result = await this.slideService.EditOnDragAndDrop(viewModel.FirstId, viewModel.SecondId);
+
+            if (result is false)
+            {
+                return new JsonResult(NotFound());
+            }
+
+            return new JsonResult(Ok());
+        }
+
+        [HttpDelete]
+        public async Task<JsonResult> Remove([FromQuery] int id)
+        {
+            var slide = await this.slideService.GetById(id);
+
+            if (slide is null)
+            {
+                return new JsonResult(NotFound());
+            }
+
+            var imagePath = Path.Combine(hostEnvironment.WebRootPath, slide.Image.TrimStart('\\'));
+            if (System.IO.File.Exists(imagePath) && !imagePath.Contains("default-background"))
+            {
+                System.IO.File.Delete(imagePath);
+            }
+
+            await this.slideService.Remove(id);
+
+
+            return new JsonResult(Ok());
+        }
+
     }
 }
